refactor: replace int with long for timestamps to avoid useless conversions

+++ snail-race-bet/java/src/main/java/snail/race/kata/domain/Bet.java
@@ -3,7 +3,7 @@ package snail.race.kata.domain;
 public record Bet(
         String gambler,
         PodiumPronostic pronostic,
-        int timestamp
+        long timestamp
 ){
 
     public boolean isInTimeFor(RaceResultProvider.SnailRace race) {
